requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
eventlet==0.36.1
werkzeug==3.0.3
python-dotenv==1.0.1
gunicorn==22.0.0
setuptools
wheel
psycopg2-binary
Flask-Compress==1.15
pywebpush
cryptography
openai
openpyxl
